+++ apps/web/module/BlogGuide/Guide/TravelBlog/TravelImages.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from "react"
 import ImageGallery from "@/module/Listing/property/image-gallery"
 import { Typography } from "@/common/components/ui/Typography"
-import ShareSave from "@/module/Listing/property/share-save"
+import ShareSave from "@/module/Listing/share-save"
 import ImageGalleryModal from "@/module/Listing/property/modals/image-gallery-modal"
 
 type T_Props = {
