+++ apps/web/module/Listing/activity/hero.tsx
@@ -1,6 +1,6 @@
 "use client"
 import { Typography } from "@/common/components/ui/Typography"
-import ShareSave from "@/module/Listing/property/share-save"
+import ShareSave from "@/module/Listing/share-save"
 import React, { useState } from "react"
 import ImageGallery from "@/module/Listing/property/image-gallery"
 import { T_Photo } from "@repo/contract-2/photos"
