+++ apps/web/module/Listing/modals/share-modal.tsx
@@ -1,6 +1,7 @@
 import ModalContainer from "@/common/components/ModalContainer"
 import { Button } from "@/common/components/ui/Button"
 import { Typography } from "@/common/components/ui/Typography"
+import { FACEBOOK_CLIENT_ID, WEB_URL } from "@/common/constants/ev"
 import {
   LucideCopy,
   LucideFacebook,
@@ -10,14 +11,14 @@ import {
   LucideMessageSquareCode,
   LucideTwitter,
 } from "lucide-react"
-import { useRouter } from "next/navigation"
+import toast from "react-hot-toast"
 
 interface SharePlaceModalProps {
   isOpen: boolean
   onClose: () => void
 }
 const SharePlaceModal = ({ isOpen, onClose }: SharePlaceModalProps) => {
-  const router = useRouter()
+  const currentLink = window.location.href;
   return (
     <ModalContainer
       isOpen={isOpen}
@@ -25,14 +26,18 @@ const SharePlaceModal = ({ isOpen, onClose }: SharePlaceModalProps) => {
       title="Share this place"
       size="sm"
     >
-      <div className="px-12 py-6">
-        <div className="flex flex-col md:flex-row gap-8 md:gap-8 pb-12">
+      <div className="px-5 py-2">
+        <div className="flex flex-col md:flex-row gap-4 pb-4">
           <div className="flex-1 md:w-1/2 2xl:w-full">
             <div className="mt-4">
               <Button
                 size={"lg"}
                 variant={"shaded"}
-                className="py-8  w-full space-x-2 pl-5 justify-start"
+                className="py-8 w-full space-x-2 pl-5 justify-start"
+                onClick={() => {
+                  toast.success("Link copied")
+                  navigator.clipboard.writeText(currentLink)
+                }}
               >
                 <LucideCopy size={"30px"} />
                 <Typography fontWeight={"semibold"}>Copy Link</Typography>
@@ -43,7 +48,6 @@ const SharePlaceModal = ({ isOpen, onClose }: SharePlaceModalProps) => {
                 size={"lg"}
                 variant={"shaded"}
                 className="py-8  w-full space-x-2 pl-5 justify-start"
-                onClick={() => (window.location.href = "sms:09760624830")}
               >
                 <LucideMessageCircle size={"30px"} />
                 <Typography fontWeight={"semibold"}>Messages</Typography>
@@ -54,10 +58,7 @@ const SharePlaceModal = ({ isOpen, onClose }: SharePlaceModalProps) => {
                 size={"lg"}
                 variant={"shaded"}
                 className="py-8  w-full space-x-2 pl-5 justify-start"
-                onClick={() =>
-                  router.push(
-                    "https://www.facebook.com/dialog/send?app_id=138566025676&link=https%3A%2F%2Fwww.airbnb.com%2Frooms%2F37250886%3Fcheck_in%3D2024-03-16%26check_out%3D2024-03-21%26guests%3D1%26adults%3D1%26s%3D25%26unique_share_id%3D77e2b1e9-f625-466f-a85f-864d355f4e70&redirect_uri=https%3A%2F%2Fwww.airbnb.com%2Frooms%2F37250886%3Fcheck_in%3D2024-03-16%26check_out%3D2024-03-21%26guests%3D1%26adults%3D1%26s%3D25%26unique_share_id%3D77e2b1e9-f625-466f-a85f-864d355f4e70"
-                  )
+                onClick={() => window.open(`https://www.facebook.com/dialog/send?app_id=${FACEBOOK_CLIENT_ID}&link=${encodeURIComponent(currentLink)}&redirect_uri=${encodeURIComponent(WEB_URL)}`, '_blank')
                 }
               >
                 <LucideMessageCircleReply size={"30px"} />
@@ -70,9 +71,7 @@ const SharePlaceModal = ({ isOpen, onClose }: SharePlaceModalProps) => {
                 variant={"shaded"}
                 className="py-8  w-full space-x-2 pl-5 justify-start"
                 onClick={() =>
-                  router.push(
-                    "https://twitter.com/intent/post?source=tweetbutton&text=&url=https%3A%2F%2Fwww.airbnb.com%2Frooms%2F37250886%3Fcheck_in%3D2024-03-16%26check_out%3D2024-03-21%26guests%3D1%26adults%3D1%26s%3D5%26unique_share_id%3D4657d164-902d-4b35-90b2-4e0e445ba0de&related=airbnb"
-                  )
+                  window.open(`https://twitter.com/intent/post?url=${encodeURIComponent(currentLink)}`, '_blank')
                 }
               >
                 <LucideTwitter size={"30px"} />
@@ -87,8 +86,7 @@ const SharePlaceModal = ({ isOpen, onClose }: SharePlaceModalProps) => {
                 variant={"shaded"}
                 className="py-8  w-full space-x-2 pl-5 justify-start"
                 onClick={() =>
-                  (window.location.href =
-                    "mailto:recipient@example.com?subject=Hello&body=How are you?")
+                  window.location.href = `mailto:?subject=${encodeURIComponent("ExploreSiargao Listing")}&body=${currentLink}`
                 }
               >
                 <LucideMail size={"30px"} />
@@ -100,7 +98,7 @@ const SharePlaceModal = ({ isOpen, onClose }: SharePlaceModalProps) => {
                 size={"lg"}
                 variant={"shaded"}
                 className="py-8  w-full space-x-2 pl-5 justify-start"
-                onClick={() => router.push("https://web.whatsapp.com/")}
+                onClick={() => window.open(`https://wa.me/?text=${currentLink}`, '_blank')}
               >
                 <LucideMessageSquareCode size={"30px"} />
                 <Typography fontWeight={"semibold"}>WhatsApp</Typography>
@@ -112,9 +110,7 @@ const SharePlaceModal = ({ isOpen, onClose }: SharePlaceModalProps) => {
                 variant={"shaded"}
                 className="py-8  w-full space-x-2 pl-5 justify-start"
                 onClick={() =>
-                  router.push(
-                    "https://www.facebook.com/sharer.php?u=https%3A%2F%2Fwww.airbnb.com%2Frooms%2F37250886%3Fcheck_in%3D2024-03-16%26check_out%3D2024-03-21%26guests%3D1%26adults%3D1%26s%3D4%26unique_share_id%3D807225a8-4974-467c-b64c-d2ed55be78e2&quote="
-                  )
+                  window.open(`https://www.facebook.com/sharer.php?u=${currentLink}`, '_blank')
                 }
               >
                 <LucideFacebook size={"30px"} />
